Cache accessible methods by name
This adds a little size and overhead to the boot process for each
class (hard to quantify) but improves the performance of
reflective accesses like JavaProxy#java_method. In JRuby 9.2.9,
those reflective accesses went directly to java.lang.Class to
look up the method, which results in extra overhead due to Class
returning copies and clones of its internal structures for
reflecting methods, and also due to our being forced to scan all
methods to look for accessible ones. JRuby 9.2.10 reused the cache
of accessible methods, reducing that overhead, but doing a linear
scan of all accessible methods to find the matching name + arg
types. This change improves the accessible methods cache by
partitioning them by name, greatly reducing the amount of scanning
needed (and requiring no scanning for methods with no overloads).

Overhead is added in the following places:

* Preparing the cached PartitionedMethods object, since it now
  caches objects in a Map<String,List> rather than just Method[].
  However this is done only once per class.
* Preparing the Ruby proxy type based on those partitioned
  methods, since the iteration now must walk two levels of
  collections. This is also done only once per nominal type.

These will add overhead to the initial preparation of a given
Java class, which might impact startup.