Dual-layer URL classloader to allow closing
Closing the URL classloader is the most hygenic way to tidy up,
but due to #6269 we must avoid doing so for jar files already on
the filesystem. We fixed this previously by simply leaving the
classloader open and attempting to manually close all temporary
jar files it may have encountered. This seems to be insufficient,
since at least on Windows those jar files are being opened with
a different stream type that does not expose the original jar
file.

This patch uses a different mechanism for closing those jar files,
injecting a non-closing URLClassLoader into the hierarchy for
uncloseable files and only closing the child URLClassLoader which
does contain jar connections that should be closed.

This was also an attempt to fix #3657 but it does not appear to
help yet.