Properly extract the 'perm' kwarg to File.open
The logic here was incorrectly ported. The original code checks
the vperm pointer passed in to know whether it can be
dereferenced. The broken code in JRuby ends up requiring a
a non-null positional vperm argument in order to use the kwarg,
which then immediately errors anyway.

The new code if the perm kwarg and the positional vperm are both
set, but if only the kwarg is given, it will be used for the new
file.

Fixes failures switching from our own native Tempfile to the
tempfile gem.

See #7973